Исправлены замечания

+++ home1/src/main/java/su/nepom/patterns/home1/model/Rotatable.java
@@ -4,7 +4,7 @@ package su.nepom.patterns.home1.model;
  * Интерфейс объекта, который можно повернуть
  */
 public interface Rotatable {
-    int getAngle();
-    void setAngle(int angle);
-    int getAngleVelocity();
+    double getAngle();
+    void setAngle(double angle);
+    double getAngleVelocity();
 }
