+++ home1/src/main/java/su/nepom/patterns/home1/model/RotatableManage.java
@@ -4,5 +4,5 @@ package su.nepom.patterns.home1.model;
  * Интерфейс объекта, скоростью поворота которого можно управлять
  */
 public interface RotatableManage extends Rotatable {
-    void setAngleVelocity(int speed);
+    void setAngleVelocity(double speed);
 }
